feat(cli): achieve CLI ↔ Framework API parity (ADR-0066)
- Implemented 11 missing subcommands: delete, get, update, disassociate, associations, traverse, path, probe-filtered, stats, metrics, watch.
- Added framework methods: \`incoming_associations\`, \`reset_metrics\`.
- Derived \`Serialize\` for \`FrameworkMetricsSnapshot\` and \`FrameworkStats\`.
- Refactored codebase to maintain < 500 LOC per file:
    - Extracted \`FrameworkMetrics\` to \`src/framework_metrics.rs\`.
    - Moved non-core \`Singularity\` methods to \`src/singularity_ext.rs\`.
    - Split tests into \`src/singularity_tests.rs\` and \`src/persistence_tests.rs\`.
- Addressed PR feedback:
    - \`csm update\` now rejects no-op requests.
    - \`csm probe-filtered --filter\` now supports both native and object-operator syntax (e.g. {"key":{"\$eq":"val"}}).
    - Reservoir \`reset_metrics\` now clears counters without resetting internal state vectors.
- Fixed CLI flag conflict: \`traverse --depth\` uses \`-D\` to avoid conflict with global \`-d\` (\`--database\`).
- Added comprehensive integration tests in \`tests/cli_parity.rs\`.
- Updated book documentation with a new CLI Reference page.
- Synchronized \`llms.txt\` and verified via \`scripts/validate.sh\`.
