+++ llms.txt
@@ -409,6 +409,7 @@ Created by: [cargo-llms-txt](https://github.com/masinc/cargo-llms-txt)
 ### src/persistence_ops.rs
 
 - impl Persistence
+- impl Persistence
 
 ### src/persistence_versions.rs
