+++ src/cli/commands/probe_filtered.rs
@@ -15,8 +15,10 @@ pub async fn run_probe_filtered(
     validate_concept_id(&args.concept_id)?;
     validate_top_k(args.top_k)?;
 
-    let filter: MetadataFilter = serde_json::from_str(&args.filter)
+    let filter_val: serde_json::Value = serde_json::from_str(&args.filter)
         .map_err(|e| CliError::Validation(format!("invalid filter JSON: {e}")))?;
+
+    let filter = parse_metadata_filter(filter_val)?;
 
     let framework = create_framework(db_path).await?;
     let concept = framework
@@ -56,4 +58,39 @@ pub async fn run_probe_filtered(
         }
     }
     Ok(())
+}
+
+fn parse_metadata_filter(v: serde_json::Value) -> Result<MetadataFilter> {
+    // Try native format first
+    if let Ok(f) = serde_json::from_value::<MetadataFilter>(v.clone()) {
+        return Ok(f);
+    }
+
+    // Try documented format: {"key": {"$eq": "val"}}
+    if let Some(obj) = v.as_object() {
+        if obj.len() == 1 {
+            let (key, op_val) = obj.iter().next().unwrap();
+            if let Some(op_obj) = op_val.as_object() {
+                if op_obj.len() == 1 {
+                    let (op, val) = op_obj.iter().next().unwrap();
+                    match op.as_str() {
+                        "$eq" => return Ok(MetadataFilter::Eq(key.clone(), val.clone())),
+                        "$in" => {
+                            if let Some(arr) = val.as_array() {
+                                return Ok(MetadataFilter::In(key.clone(), arr.clone()));
+                            }
+                        }
+                        "$exists" => {
+                            if val.as_bool().unwrap_or(false) {
+                                return Ok(MetadataFilter::Exists(key.clone()));
+                            }
+                        }
+                        _ => {}
+                    }
+                }
+            }
+        }
+    }
+
+    Err(CliError::Validation("Unsupported filter format. Use either native enum format or documented object-operator syntax.".to_string()))
 }
